+++ .github/workflows/build.yml
@@ -65,7 +65,7 @@ jobs:
             sleep 300
             export JAVA_HOME=`/usr/libexec/java_home -v 11`
             ./gradlew uiTest --info
-      - name: Archive logs
+      - name: Save screenshot
         uses: actions/upload-artifact@v2
         with:
           name: hierarchy-report
